Adding error and exception handling to Router
Calling the error method within the router will cause it to look for a
corresponding error controller. if this is found it will run it, if not it
will fall back to the internal error message.

There is a general exception handler that will call the error method with an
argument of 'exception' and also pass the exception object as an argument.

object encapsulation is now used with the output of all the controllers so
that partial output is not returned in the event of an error.